backend/requirements.txt
# GEO-AUDITOR AI - Backend Dependencies
# FastAPI & Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0

# HTTP & Scraping
playwright>=1.41.0
httpx>=0.26.0
beautifulsoup4>=4.12.0
lxml>=5.1.0

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# NLP (for future detectors)
# spacy>=3.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Utilities
python-multipart>=0.0.6

# LLM Integration
openai>=1.0.0
google-generativeai>=0.3.0
